split styles for footerstatus into single file
